cleanup splitting
